phantom_trader_project/requirements.txt
# Phantom Trader — dependencies
# Python 3.11+

# Networking
httpx>=0.24.0
websockets>=12.0

# DB (async required — prevent blocking the asyncio event loop)
aiosqlite>=0.19.0

# Telegram
python-telegram-bot>=20.0

# System monitoring
psutil>=5.9.0

# Environment variables
python-dotenv>=1.0.0
